+++ components/chatbot/Chatbot.tsx
@@ -213,7 +213,12 @@ export default function Chatbot() {
 
     {
       id: "food1",
-      message: "아직 준비중...",
+      component:
+      <CustomComponent>
+<Chatgpt></Chatgpt>
+
+      </CustomComponent>
+      
     },
     {
       id: "budget1",
